requirements.txt
Flask==2.3.2
tensorflow==2.10.0 # Gunakan versi yang kompatibel dengan model Anda
Pillow==9.2.0      # Pastikan ini kompatibel dengan TensorFlow Anda
numpy==1.23.5      # Pastikan ini kompatibel dengan TensorFlow Anda
scikit-learn==1.3.0 # Jika Anda membutuhkan fungsi-fungsi seperti label encoder
requests==2.28.1   # Untuk mengunduh file